bpf: Handle fragments in SNAT flows
The code that extracts ports in snat_v4_nat and snat_v4_rev_nat doesn't
take into account that the packet may be a second or further fragment.
In this case, garbage will be read instead of the real port numbers. Fix
this by using the existing ipv4_ct_extract_l4_ports that takes into
account fragmentation.

This change makes the behavior of the switch inside the aforementioned
functions closer to ct_extract_ports4.

Consider fragments in snat_v4_rewrite_ingress, avoid rewriting ports if
it's not the first fragment.

Fixes: 14a653ad4aac ("bpf/nat: introduce snat_v4_nat() and snat_v4_rev_nat functions")
Signed-off-by: Maxim Mikityanskiy <maxim@isovalent.com>