envoy: Update to pick up deny policy support
Ingress policy enforcement for Cilium Ingress was recently added (#28126).
Update cilium-envoy to a version that enforces also deny policies. The
only functional change in Cilium agent itself is the setting of the new
Deny flag for PortNetworkPolicyRules in pkg/xds_server.go.

The updated cilium-envoy version also adds support for overlapping port
ranges and hence also for overlapping port numbers, as in when using a
named port that resolves to the same port number used as a number
elsewhere in the policy.

Fixes: #27816
Fixes: #28126

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>