Add best-effort mode for XDP attachments
The default behavior for enabling XDP support is to enable XDP on all
interfaces. If an interface's driver does not support XDP then Cilium
will bail. This change introduces a new mode, `best-effort`, which will
swallows and logs the error when a device is encountered that does not
support XDP.

Signed-off-by: Patrick O’Brien <patrick.obrien@thetradedesk.com>