iptables: Avoid depending from DaemonConfig directly
iptables has been modularized into a cell, but still depends directly on
option.DaemonConfig global variables and helpers.

The configuration flags that are "private" to the cell are provided
through cell.Config.
Regarding the flags and the helpers shared with other parts of the code,
a SharedConfig struct is provided in the cell itself.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>