daemon: Split out `removeOldRouterState` from `restoreCiliumHostIPs`
The `restoreCiliumHostIPs` function does two things: Determine which IP
to restore, and then remove any IPs from the cilium_host device which
are not the restored one.

This commit splits out the second part, so it can be retried
independently. There is no reason to retry the first part, as it is
deterministic.

This commit prepares the code for further changes following in
subsequent commits.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>