daemon: Simplify `cilium_host` IP restoration
Before this commit, the `cilium_host` (aka router) IP restoration was
done in two steps:

The first step collected previous IPs from the filesystem and Kubernetes
and attempted to guess which one might still be valid in the new IPAM
configuration. This validation was done using either the Pod CIDR, the
VPC CIDR, or the native routing CIDR.

In a second step, the IP was then actually re-allocated via the IPAM
subsystem. That second step could however still fail - just because an
IP is part of the VPC CIDR or native routing CIDR does not mean it
actually still assigned to the node.

This commit attempts to simplify the logic: Instead of trying to guess
if the restore IP(s) could be allocated using IPAM, we just try to
allocate them. If the allocation fails, we still have a fall back logic,
and the old IP address(es) are still removed from the `cilium_host`
interface. This way, the old CIDR check becomes obsolete, as we now use
the IPAM subsystem as the source of truth if an IP can be restored or
not.

This also fixes a bug where Multi-Pool IPAM always reallocated the
router IP, because it did not implement the CIDR check in the first
step.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>