gateway-api: Bump the version to v1.0.0
There is one small change in Gateway status in upstream, in which
HTTPRoute having ResolveRef as False is still allowed to be attached
into Gateway resource, but respective Gateway listener Programmed status
should be set as False.

Due to the bug in conformance test in v1.0.0, we need to use the commit
hash from upstream main branch.

Signed-off-by: Tam Mach <tam.mach@cilium.io>