etcd: rework lease manager to additionally create sessions
Currently, the etcd lease manager is implemented directly leveraging
the Grant() and Keepalive() primitives. Yet, certain functionalities
(i.e., etcd mutexes) depend on the composite Session type. Let's
rework the manager to additionally construct such session objects,
so that in a subsequent commit we can reuse the same logic to replace
the lock session handling. Most notably, making the acquisition of said
session lazy, so that the additional overhead (given the short TTL)
can be avoided when not necessary (e.g., in the clustermesh case, as
agents access remote etcd instances in read-only mode).

The overall acquisition logic is not modified, except for the necessary
adaptations. Unit tests are updated to account for the need to propagate
a full (mocked) client object, instead of the mocked sub-interface only.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>