daemon: move ipcache restoration over to new ipcache APIs
Now that we can withhold and request specific numeric identities, we can
transition the ipcache restoration logic over. This change makes it so
that prefixes request whatever previous numeric identity they had on
restart. Once the watchers are synchronized, the ipcache will proceed
with label injection, and the state should exactly match that what was
before.

Additionally, update some tests that mimicked the now-removed restoration
logic.

Signed-off-by: Casey Callendrello <cdc@isovalent.com>