k8s ingress & gateway api: qualify envoy clusters and their references
Currently, the Clusters resources in the CiliumEnvoyConfig aren't qualified
with the namespace and name of the CEC itself.

This leads to issues when updating the CiliumEnvoyConfigs due to changes in
the K8s Ingress & Gateway API resources. If multiple resources are referencing
the same K8s Service, the Cluster resource gets deleted if one of these K8s
Ingress or Gateway resources gets deleted - and therefore breaks the other
resources.

With this commit, the name of the Cluster resource and their references gets
qualified like the rest of the resources. For the EDS lookup of the endpoint addresses,
the field `service_name` of the Cluster is used.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>