operator: introduce controller-runtime cell
Currently the Gateway API controller (and its reconcilers)
is the only one that is making us of the controller-runtime library.

To prepare its use for other use-cases, this commit extracts the
controller-runtime integration into its own cell by providing the
relevant components (Manager & Scheme) to dependent cells.

The new controller-runtime cell is responsible to start the manager itself.

Dependent cells add their reconcilers to the manager and types to the scheme
respectively.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>