ipsec/daemon: pull ipsec setup from dmn startup
Introduces the IPsec key custodian which encompasses the small part of
IPsec initialization logic which was in daemon, with the goal of
preparing MTU modularization. MTU depends on knowing the IPsec key
size (though, notably, it seemingly assumes the size doesn't change),
hence we start here with the modularisation.

This doesn't modularise a significant part of the IPsec subsystem, but
it's a incremental step in the direction. It would be conceivable to
also pull the key file watcher into the custodian, but I've opted for
the smallest diff here.

In terms of initialization ordering, this puts the IPsec key setup
_before_ the daemon startup, but after the local node store. As a
result, the SPI is guaranteed to be set earlier than before this patch.

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>