datapath/devices: Fix selection on route changes and orphanRoutes check
The addition of the check to see if routes exists without corresponding
device to the devices controller tests caused it to stop the test too
early before the test case had passed.

As the tests were faulty they didn't catch the bug where the 'Selected'
status wasn't rechecked and thus the 'veth-with-default-route' test
passed when it shouldn't have. Fix this by always checking if selected
status updates.

Fixes: 60fd85aefd ("devices: Flush routes on device delete")
Signed-off-by: Jussi Maki <jussi@isovalent.com>