api: Allow custom middleware via Hive
This commit extends the Cilium API server cell with the ability to
inject a global middleware. The middleware injected here is a global
executed after the middlewares injected by `setupGlobalMiddleware`, but
before the API method dispatch. See [1] for more details.

This allows the Hive-injected middleware to benefit from global
middleware handlers such as our panic and metrics handlers, but before
any swagger specific handlers are called. The middleware is optional, as
it is intended to be used for out-of-tree API implementations. The
middleware parameter is named, so it only applies to a particular
OpenAPI spec, thereby allowing for specialization of that API.

[1] https://goswagger.io/use/middleware.html

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>