Makefile: Move SWAGGER vars to Makefile.def
This allows other Makefiles in the tree to pull in those definitions if
they also want to invoke Swagger. There is one functional change too:
Instead of using CURDIR, we're now using ROOT_DIR in the container
volume mount. This has no impact for the existing code (where CURDIR ==
ROOT_DIR), but allows Makefiles in subdirectories to mount the full Go
module (instead of just their subdirectory), which is needed for Swagger
code gen.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>