lbipam: Add logic for requesting multiple ips with a sharing key
Updated the service compatibility check. Made the namespace check non
conditional since its a security check. Rewrote some of the logic to
be more consice. And added comments and rational to certain checks
to remind us of why we do them.

After the addition of the shared IP logic, some functions became to
large. This commit splits the code into smaller functions so the overall
flow is easier to follow.

When 2 services have a sharing key and the second service would update
so it is now incompatible with the other service it would keep the same
IP address due to a lack of compatibility checking during updates and
a logic bug in the ingress IP import code.

Add a test for sharing keys. Add logic for stacking multiple ips for one
sharing key. Add lbipam support for sharing keys.

Co-authored-by: Dylan Reimerink <dylan.reimerink@isovalent.com>
Signed-off-by: usiegl00 <50933431+usiegl00@users.noreply.github.com>