ipcache listener: drop garbage collection logic
Crafting a dedicated unit test revealed that this logic was broken,
because we considered an entry for deletion if and only if it the
given prefix was not found in the ipcache struct, and the source
of the corresponding identity was either kvstore or local. But
when the prefix is not found, we retrieve the default value of
the identity type, which will never match the source filter.

Now, considering that:

* the ipcache map gets recreated from scratch at every agent restart,
* the garbage collection logic was triggered only when either Cilium
  was configured in kvstore mode or was connected to at least one
  remote cluster,
* the etcd watcher already implements a dedicated logic to replay
  possibly missed deletion events both after relist and in case of
  complete reconnection,
* fixing this logic would likely expose us to even more headaches,
  given that we may remove entries when the ipcache is not yet
  fully synchronized from the other sources, causing the removal
  of valid entries and the disruption of existing connections,

let's just remove it.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>