ingress: implement watch for shared ciliumenvoyconfig
Ingress reconciliation needs to be triggered whenever there's a
change on the shared Ingress CiliumEnvoyConfig in the namespace
of the Cilium installation.

Therefore, this commit configures a respective watch.

To prevent an unnecessaray reconciliation of all shared Ingresses
(as they all would try to re-write the same shared CiliumEnvoyConfig),
only a pseudo ingress gets enqueued. This way all cases should be covered
(e.g. manual deletion of the shared CilumEnvoyConfig).

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>