envoy: Better track updates when qualifying Envoy resource names
Pass a boolean "updated" pointer to api.ResourceQualifiedName() in order
to better track if updates were done or not.

In the case of Http Connection Manager RouteConfig the need for the
update was ignored, so this is a bug fix for that. This omission hid a
bug where HCM RouteConfigs in our examples (and tests) refer to
"envoy-admin" cluster while unqualified cluster names should no longer be
accessible to CEC/CCEC. Fix this by renaming "envoy-admin" as
"/envoy-admin". This requres all references to "envoy-admin" cluster to
be changed to "/envoy-admin", including in any CEC/CCEC users may have.

Factor out the TCP proxy qualification logic to
qualifyTcpProxyResourceNames() which is similar to
qualifyRouteConfigurationResourceNames() to keep the main logic simpler.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>