service cache: directly observe local node store
The service cache needs to be informed about the labels associated with
the local node, and react to possible changes, in case service topology
aware hints support is enabled.

As a preparatory change to subsequently remove the local node chain and
uniform the handling of local node updates, let's switch the service
cache to directly depend on the LocalNodeStore, and observe it.

While being at it, let's also extract the associated configuration
option, and register it through hive.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>