endpoint manager: directly observe local node store
The endpoint manager needs to be informed about the labels associated
with the local node, and react to possible changes, as they get
associated with the host endpoint. To this end, it is currently
registered as a subscriber to the local node chain.

This commit switches it to directly depend on, and observe the
LocalNodeStore, to uniform the handling of local node updates and
prepare for the subsequent removal of the local node chain.
Additionally, it also drops the extraneous updates of local node
labels and annotations, which have been migrated in a previous
commit to be handled by the dedicated module.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>