datapath: Avoid Upserting AWS SG vlans
When cilium is not in charge of routing traffic,
there is no ability to avoid creating the route (`InstallEndointRoute`) while still
configuring the BPF programs (`RequireEgressProg`). This PR implements a config option
to do just that, meaning we allow the host CNI to configure routing but still enable
network policy for those endpoints. This is defaulted to enabled for aws-cni chaining.

Fixes: #27152
Signed-off-by: Alex Waring <ajmwaring@gmail.com>
Co-authored-by: Joe Stringer <joe@cilium.io>
Signed-off-by: Joe Stringer <joe@cilium.io>