probes: rename header generation test input string
This commit renames the test input for the feature header generation
test from HAVE_FIB_LOOKUP to HAVE_FEATURE_MACRO.

This is in prepation of a following commit where we remove the probing
for HAVE_FIB_LOOKUP as it is no longer needed. HAVE_FEATURE_MACRO is
test-specific and will not show up when grepping for other feature
macros that are in use in the code base.

Signed-off-by: Robin Gögge <r.goegge@isovalent.com>