bpf: nat: use rewrite helper for snat_v4_icmp_rewrite_egress_embedded()
When a local endpoint sends an ICMP_FRAG_NEEDED message, it contains some
inner packet that was originally addressed to the endpoint. The endpoint's
address/port in this inner packet potentially needs to be SNATed again, to
return it to its pre-revSNAT state.

snat_v4_icmp_rewrite_egress_embedded() handles the .daddr / .dport rewrite
of such an inner packet, based on an SNAT entry. Replace it with the
common snat_v4_rewrite_headers() helper.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>