ci: don't write github commit status on push event
Currently, some GitHub workflows are reporting back their status via
GitHub Commit Status. The main reason is to link the status of the
workflow run to a specific commit SHA. This is necessary for workflows
that are triggered via Ariane (trigger `workflows_dispatch`).

However, if the same workflow is triggered via `push` (e.g. after a PR being
merged), reporting the status via commit status is unnecessary, as the workflow
is already linked to the correct commit SHA.

Therefore, this commit suppresses the commit status reprting in case of a
`push` trigger.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>