statedb: Allow non-terminated keys
Previously statedb required the indexers to either terminate or return fixed-size
keys in order to exactly match with just prefix searching the trees. This is a
fairly large footgun.

This commit adds checks to the iterators to make sure the key matches exactly and is
not a longer key that shares the prefix. For indexers that don't terminate the key,
this may result in wasting bit of time by traversing non-matching nodes.

This was tested by "breaking" index.String by removing termination and validating
that the regression test passes.

Signed-off-by: Jussi Maki <jussi@isovalent.com>