ipsec: convert goroutines to hive jobs
Instead of launching raw goroutines, we can employ the hive job
mechanism. As is, this doesn't give an immediate benefit (except for
aligning with project direction), but enables health reporting in a
follow-up commit.

The (minor) functional difference is that instead of waiting for a
minute in between the stale key reclamation, we now start the job every
minute, as long as it completes within that minute.

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>