bgpv1: Use kube-system namespace by default for MD5 secret
Currently, BGP Control Plane uses cilium-bgp-secrets namespace by
default to fetch MD5 password authentication. Also, if not specified, it
creates a new namespace by default.

However, the rest of the cilium-related secrets such as cilium-ca,
hubble certificates, or clustermesh certificates are in kube-system
namespace. So, having a new namespace is inconsistent with rest of the
system, and also inconvenient since users need to manage one more
namespace.

This commit addresses the issue by changing the default value of
`bgpControlPlane.secretNamespace.name` to `kube-system` and
`bgpControlPlane.secretNamespace.create` to `false`. Users can always
change the namespace if they wish to have an isolation.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>