bpf: prevent clang from reordering address calculations before NULL check
Compiling the BPF complexity tests currently bails out in
tail_nodeport_nat_egress_ipv4 and tail_nodeport_nat_egress_ipv6. The error
for the v4 case is the following:

    ; return map_lookup_elem(map, tuple);
    169: (18) r1 = 0xffff9dfb4d0f9800     ; R1_w=map_ptr(off=0,ks=14,vs=40,imm=0)
    171: (85) call bpf_map_lookup_elem#1
    172: (bf) r5 = r0
    173: (07) r5 += 32
    R5 pointer arithmetic on map_value_or_null prohibited, null-check it first

The problem is easier to spot when looking at the disassembly of the function:

    ;       return map_lookup_elem(map, tuple);
    169:       r1 = 0x0 ll
    171:       call 0x1
    172:       r5 = r0
    173:       r5 += 0x20
    174:       r3 = r0
    175:       r3 += 0x24
    ;       if (*state)
    176:       if r0 != 0x0 goto +0x6da <LBB9_167>

It seems like clang has decided that it can hoist the address calculation based
on r0 before checking that r0 != 0.

Work around this issue by inserting a barrier_data() call.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>