envoy: periodic version-check with hive timer job
With the introduction of the Envoy DaemonSet mode, the Envoy version
check has changed from failing the Cilium Agent at startup to a check
that write an error message in case of a version mismatch (or error
retrieving the Envoy version).

The reason was that the Envoy DaemonSet Pod might not be ready when the
Cilium Agent performs the version check.

This commit introduces a hive timer job that periodically checks the
Envoy version (every 5min). This has the advantage that the version
gets checked periodically (in case the daemonset gets deployed with a
different version). In addition, any error gets reported via Cilium
health reporter.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>