daemon: add a device reloader that isn't as racy
With the introduction of the NodeAddress table, and the reimplementation
of the NodeAddressing on top of it, we have introduced a race in the
current, experimental runtime device detection. It's possible that the
datapath is reloaded with a new set of devices which is not reflected in
the node addressing interface yet.

We want to get rid of device manager anyway, so use this opportunity to
replace this functionality with a device reloader (this patch), and kill
the racy Listen (next patch).

Fixes: 9d7e452c20 (datapath: Reimplement NodeAddressing around Table[NodeAddress])

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>