ci-clustermesh-upgrade: Increment timeout between rollouts to 5min
Currently, the ClusterMesh upgrade test sets an explicit timeout of 1min
to wait for the Cilium Agent DaemonSet to become ready between the rollouts.

In some cases, the Pods aren't ready after 1min. Therefore, this commit
increases the timeout to 5min.

I think the most important part is that we set an explicit timeout on the
command `kubectl rollout status` - as the default is wait forever.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>