gha: add step to ensure presence/absence of the AWS iptables chains
Recently, a regression of the poststart-eni.sh script got unnoticed,
as we are not explicitly testing it as part of the E2E pipelines.

Let's add a step to the Conformance EKS and Conformance AWS-CNI
workflows to assert that, in the former case, the stale AWS iptables
chains are removed, and in the latter they are not modified.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>