pkg/lbipam: Add replacement non-time based tests
This commit adds replacements for all existing tests while the old ones
are still in place. The new tests do not start hive or the actual
LB-IPAM lifecycle, but just calls into the event handling functions.

The fixture fakes the functionality of the kube-apiserver and we inspect
the state of the LB-IPAM object after certain event. Everything is
synchoronous so there is no need for timers and assumptions about when
things are done.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>