plugins/cilium-cni: Introduce endpoint customization
This commit adds the new concept of an `EndpointConfigurator` to the CNI
plugin. An `EndpointConfigurator` allows out-of-tree implementations of
the CNI plugin to control how endpoints are created during the CNI ADD
invocation. The default implementation provided in this commit contains
no functional changes.

Tip: When reviewing this on GitHub, make sure to check "Hide whitespace"
in the diff view.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>